+++ cuhacking/lib/main.dart
@@ -66,50 +66,6 @@ class FirstScreen extends StatelessWidget {
             'user-top-read, '
             'playlist-modify-public,user-read-currently-playing');
 
-    var res = getUsersTracks();
     Navigator.pushNamed(context, '/second');
-  }
-
-  Future<void> getUsersTracks() async {
-    headers = {
-      'Content-Type': 'application/json',
-      'Accept': 'application/json',
-      'Authorization': 'Bearer $authToken'
-    };
-    var getTopSongs = await http.get(
-        'https://api.spotify.com/v1/me/top/tracks?limit=10',
-        headers: headers);
-    var topSongs = jsonDecode(getTopSongs.body);
-    var random = new Random();
-    var seedTracks =
-        topSongs["items"][random.nextInt(topSongs["items"].length)]["id"];
-
-    var getTopArtists = await http.get(
-        'https://api.spotify.com/v1/me/top/artists?time_range=medium_term&limit=10',
-        headers: headers);
-
-    var topArtists = jsonDecode(getTopArtists.body);
-
-    var seedArtists =
-        topArtists["items"][random.nextInt(topSongs["items"].length)]["id"];
-
-    var genreString = "";
-
-    for (var i = 0; i < 3; i++) {
-      var genreArray = topArtists["items"]
-          [random.nextInt(topSongs["items"].length)]["genres"];
-      genreString = genreString + genreArray[random.nextInt(genreArray.length)];
-      if (i != 2) {
-        genreString = genreString + ",";
-      }
-    }
-
-    var getRecommendations = await http.get(
-        'https://api.spotify.com/v1/recommendations?limit=25&seed_artists=$seedArtists&seed_genres=$genreString&seed_tracks=$seedTracks',
-        headers: headers);
-
-    var results = jsonDecode(getRecommendations.body);
-
-    return results["tracks"];
   }
 }
